+++ Changelog.md
@@ -1,8 +1,9 @@
 # Changelog
 
-## 0.0.2 - release
+## 0.0.2 - 2025-05-12
 
 ### Feature
+ - add test suites
  - add compare method for `ParseOptions` and `Job` struct
  - add reset function for task manager
  - add `version` command